refactor(cli): extract shared runtime resolution into cli/runtime.js
The config-dir / identity / store / realm / auth resolution (findConfigDir,
readConfig, writeConfig, resolveIdentityConfig, resolveIdentityPemPath,
resolveRealmAlias, listIdentityNames, and makeClient → openRuntime) moves out of
cli/ig.js into cli/runtime.js so the upcoming git-remote-ig helper can build a
client exactly the way the CLI does — one source of truth (DRY).

Behavior is unchanged: helpers now `throw` instead of calling die(), but every
call site runs under a top-level handler that turns a thrown Error into the same
stderr-message + exit-1 as die() did. cli/ig.js imports openRuntime as
makeClient so its call sites are untouched.

Guarded by the existing cli.test.js (55 tests) — full suite stays green (236).
